errorpage and aos added

+++ src/pages/MyToys/MyToys.jsx
@@ -20,22 +20,35 @@ const MyToys = () => {
   // console.log(users)
 
  const handleDeleteToy = _id => {
-  fetch(`http://localhost:5000/allToys/${_id}`, {
-      method: 'DELETE'
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You won't be able to revert this!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
+      fetch(`http://localhost:5000/allToys/${_id}`, {
+        method: 'DELETE'
+    })
+    .then(res => res.json())
+    .then(data =>  {
+        console.log(data)
+        if(data.deletedCount > 0){
+            Swal.fire({
+                icon: 'success',
+                title: 'Done',
+                text: 'Your Toy deleted successfully!',
+              })
+          const remaining = myToys.filter(toy => toy._id !== _id);
+          setMyToys(remaining)
+        }
+    })
+    }
   })
-  .then(res => res.json())
-  .then(data =>  {
-      console.log(data)
-      if(data.deletedCount > 0){
-          Swal.fire({
-              icon: 'success',
-              title: 'Done',
-              text: 'Your Toy deleted successfully!',
-            })
-        const remaining = myToys.filter(toy => toy._id !== _id);
-        setMyToys(remaining)
-      }
-  })
+ 
  }
